% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{catalog}
\alias{catalog}
\title{Treatment catalog}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44624 rows and 6 columns.}
\usage{
catalog
}
\description{
Treatment catalog
}
\keyword{datasets}

