% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{icd10.def}
\alias{icd10.def}
\title{ICD10 definitions according to the Institute Català de laSalut}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15447 rows and 2 columns.}
\usage{
icd10.def
}
\description{
ICD10 definitions according to the Institute Català de laSalut
}
\keyword{datasets}
