% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{icd9.def}
\alias{icd9.def}
\title{ICD9 definitions according to the Institute Català de laSalut}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22727 rows and 3 columns.}
\usage{
icd9.def
}
\description{
ICD9 definitions according to the Institute Català de laSalut
}
\keyword{datasets}

