% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population.build}
\alias{population.build}
\title{Create a population dataset from the information contained in population file and cmbd file (optional)}
\usage{
population.build(minimum_intro, maximum_intro, population_definition,
  cmbd_definition = NULL, day = NULL, other = c())
}
\arguments{
\item{population_definition}{definition of population file}

\item{cmbd_definition}{definition of cmbd file}

\item{min_intro}{minimum entrance allowed}

\item{max_intro}{maximum entrance allowed}
}
\value{
population object
}
\description{
Create a population dataset from the information contained in population file and cmbd file (optional)
}

