% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population.filter_age}
\alias{population.filter_age}
\title{Population filtered according to age filter}
\usage{
population.filter_age(population, min_age = -Inf, max_age = Inf)
}
\arguments{
\item{population}{population object}

\item{min_age}{minimum age allowed}

\item{max_age}{maximum age allowed}
}
\value{
population object
}
\description{
Population filtered according to age filter
}
