% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{problems.build}
\alias{problems.build}
\title{Read a delimited sisap file}
\usage{
problems.build(ecap_definition = NULL, cmbd_definition = NULL,
  ocip_filter = NULL)
}
\arguments{
\item{ecap_definition}{Definition of sisap ecap file}

\item{cmbd_definition}{Definition of sisap cmbd file}

\item{ocip_filter}{filtering to be applied (optional)}
}
\value{
returns a data.frame in \code{\link{tbl_df}} format
}
\description{
Read a delimited sisap file
}

