% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{problems.first_event}
\alias{problems.first_event}
\title{Get the date of the first event}
\usage{
problems.first_event(df.disease, xml_disease_file, diseases, any = TRUE,
  where_icd9 = c("dp", paste0("ds", 1:9)), where_icd10 = "ecap",
  icd9.list = NULL, icd10.list = NULL)
}
\arguments{
\item{df.disease}{data.frame containing the registry}

\item{xml_disease_file}{xml file containing the disease definition}

\item{diseases}{name of diseases of interest (must be contained in the xml file definition)}

\item{any}{logical indicating whether to calculate the overall minimum (TRUE) or the separated by disease (FALSE)}

\item{where_icd9}{character indicating where to find the information inside column field}

\item{where_icd10}{character indicating where to find the information inside column field}

\item{icd9.list}{vector with the ICD-9 codes should be considered}

\item{icd10.list}{vector with the ICD-10 codes should be considered}
}
\value{
returns a data.frame with the date of the first event
}
\description{
Get the date of the first event
}

