% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{problems.first_procedure}
\alias{problems.first_procedure}
\title{Get the date of the first procedure}
\usage{
problems.first_procedure(df.disease, xml_disease_file, procedures, any = TRUE,
  where_icd9 = c("pp", paste0("ps", 1:7)))
}
\arguments{
\item{df.disease}{data.frame containing the registry}

\item{xml_disease_file}{xml file containing the disease definition}

\item{any}{logical indicating whether to calculate the overall minimum (TRUE) or the separated by disease (FALSE)}

\item{where_icd9}{character indicating where to find the information inside column field}

\item{diseases}{name of diseases of interest (must be contained in the xml file definition)}
}
\value{
returns a data.frame with the date of the first procedure
}
\description{
Get the date of the first procedure
}

