% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisap.R
\name{sisap.file_definition}
\alias{sisap.file_definition}
\title{Sisap file definition}
\usage{
sisap.file_definition(filename, delim = "@", col_names = NULL,
  integer_vars = NULL, double_vars = NULL, logical_vars = NULL,
  date_vars = NULL, type = NULL, numcol = NULL)
}
\arguments{
\item{filename}{path to the file}

\item{delim}{file delimiter}

\item{integer_vars}{variables to be considered integers}

\item{double_vars}{variables to be considered doubles}

\item{logical_vars}{variables to be considered logicals}

\item{date_vars}{variables to be considered dates}
}
\value{
file definition
}
\description{
Sisap file definition
}

