% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisap.R
\name{sisap.read_file}
\alias{sisap.read_file}
\title{Read a delimited sisap file}
\usage{
sisap.read_file(file_def, vars = names(file_def$col_names),
  locale = readr::locale(date_format = "\%Y\%m\%d"), ...)
}
\arguments{
\item{file_def}{file definition}

\item{vars}{variables of interest}

\item{...}{other parameters passed to \code{\link{read_delim}}}
}
\value{
returns a data.frame in \code{\link{tbl_df}} format
}
\description{
Read a delimited sisap file
}

