% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{treatment.build}
\alias{treatment.build}
\title{Create a treatment data.frame from registry}
\usage{
treatment.build(treatment_definition, ocip_filter = NULL, day_of_month = 28)
}
\arguments{
\item{treatment_definition}{definition of treatment file}

\item{ocip_filter}{filtering to be applied (optional)}

\item{day_of_month}{day of month to be considered as a facturation date (default=28)}
}
\value{
a data.frame containing the facturations
}
\description{
Create a treatment data.frame from registry
}
