% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{treatment.periods}
\alias{treatment.periods}
\title{Create a treatment data.frame with a treatment from registry}
\usage{
treatment.periods(df.treatment, months_effect = 6, include_env = TRUE)
}
\arguments{
\item{df.treatment}{data.frame with treatment}

\item{months_effect}{months effect}

\item{include_env}{consider the number of containers in the effect}
}
\value{
a data.frame containing the facturations
}
\description{
Create a treatment data.frame with a treatment from registry
}

